# sdks/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "LiteGen API",
        "description": "Universal proxy for AI image and video generation.",
        "license": {
            "name": "MIT",
            "identifier": "MIT"
        },
        "version": "0.1.0"
    },
    "paths": {
...
        "/v1/auth/csrf": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "GET /v1/auth/csrf \u2014 Return the CSRF token for the current session.",
                "operationId": "csrf_token",
                "responses": {
                    "200": {
                        "description": "CSRF token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CsrfResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CsrfResponse": {
                "type": "object",
                "required": [
                    "csrf_token"
                ],
                "properties": {
                    "csrf_token": {
                        "type": "string"
                    }
                }
            },
...
            "ErrorDetail": {
                "type": "object",
                "required": [
                    "message",
                    "type"
                ],
                "properties": {
                    "code": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "message": {
                        "type": "string"
                    },
                    "provider_error": {},
                    "type": {
                        "type": "string"
                    }
                }
            },
            "ErrorResponse": {
                "type": "object",
                "required": [
                    "error"
                ],
                "properties": {
                    "error": {
                        "$ref": "#/components/schemas/ErrorDetail"
                    }
                }
            },
...
        }
    },
...
}